+++ test/index.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const { parser, Changelog, Release } = require('../src');
 const assert = require('assert');
+const Semver = require('semver');
 
 const changelog = parser(fs.readFileSync(__dirname + '/changelog.md', 'UTF-8'));
 const expected = fs.readFileSync(__dirname + '/changelog.expected.md', 'UTF-8');
@@ -16,6 +17,47 @@ describe('Changelog testing', function() {
 
     it('should match the generated changelog with the expected', function() {
         assert.equal(changelog.toString().trim(), expected.trim());
+    });
+
+    describe('findRelease', function() {
+        it('should find an Unreleased release if no argument is passed in', function() {
+            const changelog = new Changelog('Changelog');
+            const unreleased = new Release();
+            const versioned = new Release('1.2.3');
+            changelog.addRelease(unreleased).addRelease(versioned);
+            assert.equal(changelog.findRelease(), unreleased);
+        });
+
+        it('should find an Unreleased release if null is passed in', function() {
+            const changelog = new Changelog('Changelog');
+            const unreleased = new Release();
+            const versioned = new Release('1.2.3');
+            changelog.addRelease(unreleased).addRelease(versioned);
+            assert.equal(changelog.findRelease(null), unreleased);
+        });
+
+        it('should return undefined when there is no Unreleased release', function() {
+            const changelog = new Changelog('Changelog');
+            const versioned = new Release('1.2.3');
+            changelog.addRelease(versioned);
+            assert.equal(changelog.findRelease(null), undefined);
+        });
+
+        it('should find the given release', function() {
+            const changelog = new Changelog('Changelog');
+            const unreleased = new Release();
+            const versioned = new Release('1.2.3');
+            changelog.addRelease(unreleased).addRelease(versioned);
+            assert.equal(changelog.findRelease('1.2.3'), versioned);
+        });
+
+        it('should return undefined for a non-existent release', function() {
+            const changelog = new Changelog('Changelog');
+            const unreleased = new Release();
+            const versioned = new Release('1.2.3');
+            changelog.addRelease(unreleased).addRelease(versioned);
+            assert.equal(changelog.findRelease('1.0.0'), undefined);
+        });
     });
 });
 
@@ -41,6 +83,38 @@ describe('Release testing', function() {
             assert.equal(new Release().removed('removed').isEmpty(), false);
             assert.equal(new Release().fixed('fixed').isEmpty(), false);
             assert.equal(new Release().security('security').isEmpty(), false);
+        });
+    });
+
+    describe('setVersion', function() {
+        it('should update the version of a null-version release', function() {
+            const release = new Release();
+            assert.equal(release.version, undefined);
+            release.setVersion('1.2.3');
+            assert.equal(release.version instanceof Semver, true);
+            assert.equal(release.version.toString(), '1.2.3');
+        });
+
+        it('should update the version of a versioned release', function() {
+            const release = new Release('1.2.2');
+            assert.equal(release.version instanceof Semver, true);
+            assert.equal(release.version.toString(), '1.2.2');
+            release.setVersion('1.2.3');
+            assert.equal(release.version instanceof Semver, true);
+            assert.equal(release.version.toString(), '1.2.3');
+        });
+
+        it("should sort the parent changelog's releases", function() {
+            const release = new Release('1.2.2');
+            let sortCalled = false;
+            release.changelog = {
+                sortReleases() {
+                    sortCalled = true;
+                }
+            };
+            assert.equal(sortCalled, false);
+            release.setVersion('1.2.3');
+            assert.equal(sortCalled, true);
         });
     });
 });