+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.4] - 2018-09-03
+
+### Added
+
+- Added `findRelease` function to `Changelog` for finding a release by version number - [#6]
+- Added `setVersion` function to `Release` for changing the version of a release - [#6]
+
 ## [0.6.3] - 2018-08-22
 
 ### Fixed
@@ -107,11 +114,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[#6]: https://github.com/oscarotero/keep-a-changelog/issues/6
 [#5]: https://github.com/oscarotero/keep-a-changelog/issues/5
 [#4]: https://github.com/oscarotero/keep-a-changelog/issues/4
 [#3]: https://github.com/oscarotero/keep-a-changelog/issues/3
 [#1]: https://github.com/oscarotero/keep-a-changelog/issues/1
 
+[0.6.4]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.6.0...v0.6.1